zlt-status/requirements.txt
Flask==1.1.2
gevent==20.9.0
pyvim==3.0.2
redis==3.5.3
Flask_SocketIO==4.3.1
eventlet==0.29.1
easysnmp==0.2.5
pyVmomi==7.0.1
python-socketio==4.6.0
